Add files via upload
This the registration form made with HTML and CSS. It has fields like  the username field, full name field, email id which has an @ compulsory in it and the password field. All the fields are compulsory to fill out otherwise the user can't proceed. There are warnings given if the field is missed out or @ is not mentioned in the mail-id. User is provided the option to view the password. This registration form is highly responsive i.e., it is suitable for all devices.